Добавил реализацию фрагментов
