+++ app/src/main/java/ru/techpark/agregator/FragmentNavigator.java
@@ -1,0 +1,5 @@
+package com.example.dz;
+
+interface FragmentNavigator {
+    void navigateToAnotherFragment(int num, int col);
+}
